# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CEX Public & Admin API
  version: 1.0.0
  summary: REST/JSON API for the centralized crypto exchange (frozen DB v1.0).
  description: |
    Contract-first API for the exchange. Generated clients (Flutter, admin SPA)
    are the consumers; **this spec is the boundary**.

    ## Conventions
    - **Transport:** REST/JSON over HTTPS only. UTC ISO-8601 timestamps.
    - **Envelope (success):** `{ "success": true, "data": <T>, "meta"?: {...} }`
    - **Envelope (error):** `{ "success": false, "error": { "code", "message", "details"? } }`
      Clients **switch on `error.code`** (stable enum), never on prose or HTTP status alone.
    - **Money:** all monetary values are **decimal strings** (e.g. `"100.50"`), never
      JSON numbers — mirrors the `NUMERIC` ledger and avoids float drift. Pattern
      `^-?\d+(\.\d+)?$`. Scale is asset-defined (INR=2, USDT=6, …).
    - **IDs:** UUID v4 strings unless noted (`trades.seq` and log ids are int64 strings).
    - **Pagination:** cursor-based — `?cursor=&limit=` → `data.items[]` + `data.nextCursor`
      (null at end). No offset paging (unstable for append-only history).
    - **Versioning:** URL-major (`/api/v1`, `/admin/v1`). See `VERSIONING` in API_DESIGN.md.

    ## Authentication
    - **UserBearer** — short-lived access JWT (10–15 min) in `Authorization: Bearer`.
      Rotate via `POST /auth/refresh` (rotating refresh token, reuse-detected).
...

    ## Idempotency
...
    it carry `x-idempotent: required`.

...
servers:
...
security:
  - UserBearer: []

# ============================================================================
# PATHS — PUBLIC API (/api/v1)
# ============================================================================
paths:
...
  /withdrawals/{withdrawalId}:
    get:
      tags: [Withdrawals]
      operationId: getWithdrawal
      parameters: [{ $ref: '#/components/parameters/WithdrawalId' }]
      summary: Single withdrawal detail.
      x-rate-limit: { policy: standard, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/WithdrawalEnvelope' } } }, description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access JWT (10–15 min TTL).
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access JWT on the admin host; TOTP mandatory at login.

  # ---- reusable parameters ----
  parameters:
...
    WithdrawalId: { name: withdrawalId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Unauthorized:
      description: Missing/invalid/expired token.
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: UNAUTHORIZED, message: Authentication required } } } }
...
    NotFound:
      description: Resource not found (or not visible to caller).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: NOT_FOUND, message: Resource not found } } } }
...
  schemas:
...
    Decimal:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: Arbitrary-precision decimal as a string (never a float).
      example: '100.500000'
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
        details: { description: Optional structured context (e.g. field errors). }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: '#/components/schemas/Error' }
...
    ErrorCode:
      type: string
      description: Stable machine-readable error code (clients switch on this).
      enum:
        - BAD_REQUEST
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
        - IDEMPOTENCY_IN_PROGRESS
        - STEP_UP_REQUIRED
        - INVALID_CREDENTIALS
        - INVALID_TOTP
        - CHALLENGE_EXPIRED
        - INVALID_REFRESH_TOKEN
        - REFRESH_REUSE_DETECTED
        - ACCOUNT_LOCKED
        - ACCOUNT_FROZEN
        - EMAIL_NOT_VERIFIED
        - EMAIL_TAKEN
        - PHONE_TAKEN
        - TOTP_ALREADY_ENABLED
        - KYC_REQUIRED
        - KYC_IN_REVIEW
        - KYC_ALREADY_APPROVED
        - INSUFFICIENT_BALANCE
        - LIMIT_EXCEEDED
        - AMOUNT_BELOW_MINIMUM
        - WITHDRAWALS_DISABLED
        - INR_DISABLED
        - ADDRESS_NOT_WHITELISTED
        - ADDRESS_IN_COOLDOWN
        - ADDRESS_ALREADY_LISTED
        - WITHDRAWAL_NOT_CANCELLABLE
        - WITHDRAWAL_NOT_APPROVABLE
        - DUAL_CONTROL_SAME_APPROVER
        - BANK_NOT_VERIFIED
        - QUOTE_EXPIRED
        - QUOTE_NOT_FOUND
        - DUPLICATE_CLIENT_ORDER_ID
        - ORDER_NOT_CANCELLABLE
        - MARKET_HALTED
        - MARKET_EXISTS
        - TICK_SIZE_VIOLATION
        - STEP_SIZE_VIOLATION
        - MIN_NOTIONAL
        - POST_ONLY_WOULD_CROSS

    # ======================= enums (mirror DB) =======================
...
    WithdrawalStatus: { type: string, enum: [REQUESTED, RISK_CHECK, PENDING_APPROVAL, APPROVED, QUEUED, SIGNING, BROADCAST, CONFIRMING, COMPLETED, REJECTED, FAILED, CANCELLED] }
...
    Withdrawal:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        asset: { type: string }
        chain: { type: string }
        toAddress: { type: string }
        amount: { $ref: '#/components/schemas/Decimal' }
        fee: { $ref: '#/components/schemas/Decimal' }
        tdsAmount: { $ref: '#/components/schemas/Decimal' }
        netAmount: { $ref: '#/components/schemas/Decimal' }
        status: { $ref: '#/components/schemas/WithdrawalStatus' }
        txHash: { type: [string, 'null'] }
        requestedAt: { $ref: '#/components/schemas/Timestamp' }
        completedAt: { type: [string, 'null'], format: date-time }
...
    WithdrawalEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/Withdrawal' } } }
...